MeasuredTimer: Use double instead of float for elapsed time

+++ demos/AnimatedGLWidget.cpp
@@ -4,8 +4,8 @@ AnimatedGLWidget::AnimatedGLWidget(QWidget* const parent) :
 	GLWidget(parent),
 	timer(1000 / 60, this)
 {
-	connect(&timer, SIGNAL(timeout(const float&)), this, SLOT(updateGL()));
-	connect(&timer, SIGNAL(timeout(const float&)), this, SLOT(tick(const float&)));
+	connect(&timer, SIGNAL(timeout(const double&)), this, SLOT(updateGL()));
+	connect(&timer, SIGNAL(timeout(const double&)), this, SLOT(tick(const double&)));
 	timer.startOnShow(this);
 	updateOnCameraMove(false);
 }
