+++ src/Alpha.h
@@ -58,7 +58,7 @@ protected:
 	}
 
 private slots:
-	void tick(const float& elapsed);
+	void tick(const double& elapsed);
 };
 
 void drawCameraOrientedText(const gl::Camera& camera, FTFont* const font, const char* text, int zOffset);
